requirements.txt
Flask
icrawler
gunicorn
pillow
